refactor(AppUpdater): split AppUpdater.swift into three files
AppUpdater.swift was 693 lines, exceeding the project's SwiftLint
file_length limit of 620. Rather than trimming comments, split the file
at natural MARK boundaries:

- AppUpdater.swift: class declaration, stored properties, init, deinit
- AppUpdater+UpdateFlow.swift: checkAndHandle, checkForUpdate,
  rehydrateCachedUpdateIfNewer, handle (public entry-point logic)
- AppUpdater+Download.swift: downloadUpdate (download + verify + cache)

No logic changes. All three files are well under the 620-line limit.
CryptoKit import moved to AppUpdater+Download.swift where it is used.
