+++ Sources/AppUpdater/AppUpdater.swift
@@ -1,6 +1,5 @@
 // AppUpdater.swift
 // AppUpdater
-import CryptoKit
 import Foundation
 // AppKit is unavailable in the SPM headless test runner — this guard is
 // required for `swift test` even though the package is macOS(.v26)-only.
@@ -186,7 +185,7 @@ public final class AppUpdater {
     // AppKit is unavailable in the SPM headless test runner — the #if guard is
     // required for `swift test` even though the package is macOS(.v26)-only.
     //
-    // ── Why nonisolated(unsafe) var, and why it is safe ──────────────────────
+    // ── Why nonisolated(unsafe) var, and why it is safe ─────────────────────────
     //
     // `deinit` is a nonisolated context — the Swift compiler does not permit
     // accessing a `@MainActor`-isolated property from `deinit` directly.
@@ -232,7 +231,7 @@ public final class AppUpdater {
     // nonisolated access, re-evaluate whether the underlying constraint is the
     // same (deinit invalidation of a thread-safe object) before using it.
     //
-    // ── Known alternative (not pursued, low priority) ─────────────────────────
+    // ── Known alternative (not pursued, low priority) ───────────────────────
     //
     // `scheduleBackgroundCheck` assigns `activity` exactly once. It could
     // theoretically be refactored to `nonisolated(unsafe) let` — set once at
@@ -334,360 +333,5 @@ public final class AppUpdater {
         // deinit is never reached while a download is in flight. Adding a
         // cancel call here would be dead code. See the "Fire-and-forget
         // rationale" section in the downloadUpdate doc comment.
-    }
-
-    // MARK: - Convenience entry points
-
-    /// Runs a full update check and handles the result.
-    ///
-    /// On `.updateAvailable` the release is downloaded/cached via `handle`.
-    /// `.upToDate` and `.failed` are no-ops here (the background scheduler owns
-    /// the stale-row-clearing policy) — this mirrors a launch-time check that
-    /// must never clear a ready-to-install update after a transient failure.
-    public func checkAndHandle(state: any UpdateStateProviding) async {
-        let beta = betaChannelProvider()
-        switch await checkForUpdate(betaChannel: beta) {
-        case .updateAvailable(let release):
-            appUpdaterLogger.debug("update available: \(release.tagName, privacy: .public) (beta=\(beta, privacy: .public))")
-            await handle(release, state: state)
-        case .upToDate:
-            appUpdaterLogger.debug("no update available (beta=\(beta, privacy: .public))")
-        case .failed(let error):
-            appUpdaterLogger.debug("update check failed: \(String(describing: error), privacy: .public) (beta=\(beta, privacy: .public))")
-        }
-    }
-
-    /// Runs a channel-aware update check via the injected `ReleaseProvider`.
-    ///
-    /// Fetches via `provider.fetchLatestRelease` and performs the `isNewer`
-    /// comparison here in `AppUpdater`. `UpdateChecker.checkForUpdate` remains
-    /// `public` as an escape hatch for callers who need a raw `UpdateCheckResult`
-    /// without constructing an `AppUpdater` instance.
-    ///
-    /// Intentionally `internal` — `checkAndHandle` is the designed public entry
-    /// point for host apps.
-    func checkForUpdate(betaChannel: Bool) async -> UpdateCheckResult {
-        guard !currentVersion.isEmpty else {
-            return .failed(UpdateCheckError.missingVersionKey)
-        }
-        guard let latest = await provider.fetchLatestRelease(
-            repo: repo,
-            betaChannel: betaChannel,
-            assetName: assetName
-        ) else {
-            return .failed(UpdateCheckError.noReleasesFound)
-        }
-        guard UpdateChecker.isNewer(latest.tagName, than: currentVersion) else {
-            return .upToDate
-        }
-        return .updateAvailable(release: latest)
-    }
-
-    /// Rehydrates cached download state on launch, if a newer zip is cached.
-    ///
-    /// Reads this updater's scoped `UserDefaults` keys. If a cached zip still
-    /// exists on disk and its version is newer than `currentVersion`, the host
-    /// state is rehydrated and the available-update label is set so the install
-    /// affordance is visible immediately — even offline, before any network
-    /// check runs. Otherwise stale keys are cleared.
-    ///
-    /// ## Stale zip purge
-    ///
-    /// `purgeStaleZips(keeping:)` is called unconditionally at the top of this
-    /// function, before the guard, so orphaned zips from the previous session
-    /// are swept regardless of whether rehydration succeeds or fails. The
-    /// `keepURL` is derived from `UserDefaults` before the guard block so the
-    /// live zip (if any) is never deleted even when rehydration returns early.
-    ///
-    /// REVIEWER: Do NOT move the `purgeStaleZips` call inside the guard's
-    /// success branch. The orphans this sweep targets — specifically the open-n
-    /// failure path where UserDefaults are cleared before the zip is deleted —
-    /// are present precisely when the guard fails (no valid path in UserDefaults).
-    /// Moving it inside the success branch would silently miss every orphan.
-    ///
-    /// Call this before `checkAndHandle` on startup.
-    public func rehydrateCachedUpdateIfNewer(state: any UpdateStateProviding) {
-        // Derive the live zip URL from UserDefaults before the guard so
-        // purgeStaleZips knows which file to preserve regardless of which
-        // branch is taken below.
-        let liveZipURL = defaults.string(forKey: keys.cachedUpdateZipPath)
-            .map { URL(fileURLWithPath: $0) }
-
-        // Purge stale zips unconditionally. Covers:
-        // - Normal install cycles (prior version's zip left on disk)
-        // - open -n failure path (UserDefaults cleared, zip orphaned)
-        // - Any future edge case leaving zips without a UserDefaults pointer
-        // The live zip (liveZipURL) is preserved if present.
-        purgeStaleZips(keeping: liveZipURL)
-
-        guard let path = defaults.string(forKey: keys.cachedUpdateZipPath),
-              let version = defaults.string(forKey: keys.cachedUpdateVersion),
-              FileManager.default.fileExists(atPath: path),
-              UpdateChecker.isNewer(version, than: currentVersion) else {
-            // Reached when the keys were never written, the zip was deleted, or
-            // the cached version is no longer newer (already installed). In all
-            // cases clear the stale keys.
-            clearCachedDefaults()
-            return
-        }
-        state.rehydrateCachedUpdate(zipURL: URL(fileURLWithPath: path), version: version)
-        state.setAvailableUpdate(version)
-    }
-
-    // MARK: - Entry point
-
-    /// Responds to a newly discovered available release.
-    ///
-    /// 1. If a matching cached zip already exists for this version, rehydrates
-    ///    the host state and returns without re-downloading.
-    /// 2. If the release has no matching zip asset **or** no checksum sidecar
-    ///    URL, sets the host failure state (curl-install fallback) and returns.
-    ///    Both are treated identically: the release cannot be safely downloaded
-    ///    and verified, so the download path is never entered and
-    ///    `setDownloadStarted()` is never called.
-    /// 3. Otherwise starts a background download; the host state is updated on
-    ///    the main actor when it completes.
-    ///
-    /// `setAvailableUpdate` is called AFTER the `isDownloading` guard (step 3)
-    /// so the version label and the cached zip always agree.
-    public func handle(_ release: AvailableRelease, state: any UpdateStateProviding) async {
-
-        // ── 1. Already cached? ───────────────────────────────────────────────────────────────────
-        let cachedVersion = defaults.string(forKey: keys.cachedUpdateVersion)
-        let cachedPath = defaults.string(forKey: keys.cachedUpdateZipPath)
-        if let cachedVersion, cachedVersion == release.tagName, let path = cachedPath {
-            if FileManager.default.fileExists(atPath: path) {
-                state.rehydrateCachedUpdate(zipURL: URL(fileURLWithPath: path), version: cachedVersion)
-                state.setAvailableUpdate(release.tagName)
-                return
-            }
-            clearCachedDefaults()
-        }
-
-        // ── 2. Asset or checksum sidecar absent from release? ────────────────────────────────────
-        // Both conditions are treated identically: the release cannot be safely
-        // downloaded and integrity-verified, so we surface the curl-install
-        // fallback without ever entering the download path.
-        //
-        // ⚠️ The checksumURL guard MUST live here (step 2), not inside
-        // downloadUpdate(). Allowing a nil checksumURL to reach the download
-        // path would cause setDownloadStarted() to fire first — producing a
-        // visible spinner flash — before downloadUpdate() throws and eventually
-        // calls setUpdateFailed(). Catching it here keeps the UI transition
-        // direct: setAssetMissing() → curl-install fallback, no spinner shown.
-        let wantedAsset = assetName(release.tagName)
-        guard let asset = release.assets.first(where: { $0.name == wantedAsset }) else {
-            state.setAvailableUpdate(release.tagName)
-            state.setAssetMissing()
-            return
-        }
-        guard release.checksumURL != nil else {
-            state.setAvailableUpdate(release.tagName)
-            state.setAssetMissing()
-            return
-        }
-
-        // ── 3. In-flight guard ───────────────────────────────────────────────────────────────────
-        guard !isDownloading else { return }
-        isDownloading = true
-
-        state.setAvailableUpdate(release.tagName)
-
-        // ── 3b. Move to downloading state ───────────────────────────────────────────────────────
-        // clearCachedDefaults() MUST run before setDownloadStarted(). If the
-        // process is force-quit in the window between these two calls, the
-        // invariant we need to preserve is: UserDefaults must not contain a
-        // cached-zip path that points at a file which was never fully
-        // downloaded. Running clearCachedDefaults() first guarantees that a
-        // crash here leaves UserDefaults clean; a subsequent launch finds no
-        // cached path and triggers a fresh check. The reverse order would leave
-        // a stale path in defaults pointing at nothing, causing
-        // rehydrateCachedUpdateIfNewer() to silently clear it on next launch
-        // rather than offering the install — a recoverable but confusing state.
-        // Do NOT reorder these two calls or insert an `await` between them.
-        clearCachedDefaults()
-        state.setDownloadStarted()
-
-        let downloadURL = asset.browserDownloadURL
-        let checksumURL = release.checksumURL
-        let tagName = release.tagName
-
-        // REVIEWER: Do NOT store this task handle, do NOT add cancellation, and
-        // do NOT add a `downloadTask` property. Fire-and-forget is the correct
-        // pattern here — this is a deliberate design decision, not an oversight.
-        //
-        // Full rationale is in the "## Fire-and-forget rationale" section of
-        // the `downloadUpdate` doc comment below. Summary:
-        //
-        // 1. AppUpdater is owned by AppDelegate for the app's entire lifetime.
-        //    Deallocation mid-download is architecturally impossible — there is
-        //    no scenario where a stored handle + deinit cancel would ever fire.
-        //
-        // 2. `isDownloading = true` (step 3 above) already serialises all
-        //    concurrent handle() calls. A second Task is never started while
-        //    one is in flight — the guard above returns early.
-        //
-        // 3. `downloadUpdate` is a one-shot operation, not a loop. The P3
-        //    concurrency doc pattern (stored task + generation counter) applies
-        //    to long-lived polling loops, not to one-shot fire-and-forget work.
-        //
-        // 4. `@MainActor` isolation guarantees that all state callbacks inside
-        //    `downloadUpdate` (setDownloadComplete, setUpdateFailed,
-        //    isDownloading = false) are serialised — no stored handle is needed
-        //    for ordering guarantees.
-        //
-        // The task is named for Instruments / crash-log debuggability only
-        // (reach-goal principle 6, SE-0462). That is the sole purpose of
-        // Task(name:) here.
-        Task(name: "AppUpdater.download") {
-            await self.downloadUpdate(from: downloadURL, checksumURL: checksumURL, version: tagName, state: state)
-        }
-    }
-
-    // MARK: - Download
-
-    /// Downloads the zip and its SHA-256 sidecar in parallel, verifies
-    /// integrity, then caches the verified zip and updates host state.
-    ///
-    /// ## Fire-and-forget rationale
-    ///
-    /// This function is invoked from a fire-and-forget `Task(name: "AppUpdater.download")`
-    /// in `handle()`. Reviewers familiar with structured concurrency may flag
-    /// the absence of a stored task handle or cancellation support as a
-    /// defect. It is not — this is a deliberate design decision.
-    ///
-    /// **Why no stored handle / cancellation:**
-    ///
-    /// 1. **Lifetime:** `AppUpdater` is owned by `AppDelegate` and lives for
-    ///    the entire app lifetime. `deinit` is never reached while a download
-    ///    is in flight. `downloadTask?.cancel()` in `deinit` would be dead code.
-    ///
-    /// 2. **Concurrency guard:** `isDownloading = true` is set before the
-    ///    `Task` is spawned in `handle()`. Every subsequent `handle()` call
-    ///    returns immediately at the in-flight guard. There is no scenario
-    ///    where a second download `Task` is started while this one runs.
-    ///
-    /// 3. **One-shot, not a loop:** The P3 concurrency-doc pattern of storing a
-    ///    task handle applies to long-lived loops (`pollTask`, generation
-    ///    counters). A single download-verify-cache operation does not benefit
-    ///    from that pattern — there is nothing to cancel mid-flight that would
-    ///    leave state cleaner than letting the do/catch complete naturally.
-    ///
-    /// 4. **`@MainActor` serialisation:** All state mutations in this function
-    ///    (`setDownloadComplete`, `setUpdateFailed`, `isDownloading = false`)
-    ///    run on the main actor. Order and exclusivity are guaranteed by the
-    ///    actor executor without a stored handle.
-    ///
-    /// REVIEWER: Do NOT raise "unstructured task" or "missing cancellation" as
-    /// a finding on `handle()` or `downloadUpdate`. Both have been evaluated
-    /// and the fire-and-forget pattern is correct for this case. Adding a
-    /// stored `downloadTask` property would be over-engineering a problem that
-    /// does not exist in this codebase.
-    ///
-    /// ## HTTP status validation
-    ///
-    /// Both the zip response and the checksum sidecar response are validated
-    /// for HTTP 200 before their bodies are used. This matters most for the
-    /// sidecar: GitHub's CDN returns a 404 HTML page (not an error throw) when
-    /// a release was published without a `.sha256` file. Without the status
-    /// check the HTML body would be fed into the hex parser, `expectedHex`
-    /// would be `"<!DOCTYPE"`, `verifyChecksum` would throw a digest mismatch,
-    /// and Console.app would log "checksum mismatch" instead of "sidecar
-    /// returned HTTP 404" — making the failure significantly harder to diagnose.
-    /// The install is blocked correctly in both cases; this is purely a
-    /// diagnosability improvement.
-    private func downloadUpdate( // skipcq: SW-R1002 — reviewed; complexity acceptable for this download+verify flow
-        from url: URL,
-        checksumURL: URL?,
-        version: String,
-        state: any UpdateStateProviding
-    ) async {
-        var tempURL: URL?
-        do {
-            let sessionConfig = URLSessionConfiguration.ephemeral
-            sessionConfig.timeoutIntervalForRequest = 30
-            sessionConfig.timeoutIntervalForResource = 300
-            let session = URLSession(configuration: sessionConfig)
-            defer { session.finishTasksAndInvalidate() }
-
-            // Safety net: handle() now guards against nil checksumURL in step 2
-            // before entering the download path, so this branch should be
-            // unreachable in normal flow. It is kept as a last-resort defensive
-            // check — if a future refactor bypasses the step-2 guard, this
-            // prevents a nil URL from reaching URLSession.
-            guard let checksumURL else {
-                throw URLError(.resourceUnavailable)
-            }
-
-            async let zipDownload = session.download(from: url)
-            async let checksumDownload = session.data(from: checksumURL)
-            let (downloadedURL, zipResponse) = try await zipDownload
-            tempURL = downloadedURL
-            let (checksumData, checksumResponse) = try await checksumDownload
-
-            // ── Validate zip HTTP status ─────────────────────────────────────
-            guard let zipHTTP = zipResponse as? HTTPURLResponse else {
-                throw URLError(.badServerResponse)
-            }
-            guard zipHTTP.statusCode == 200 else {
-                appUpdaterLogger.error("zip download returned HTTP \(zipHTTP.statusCode, privacy: .public)")
-                throw URLError(.badServerResponse)
-            }
-
-            // ── Validate checksum sidecar HTTP status ────────────────────────
-            // A non-200 here most commonly means the release was published
-            // without a .sha256 sidecar file. Without this guard the CDN's
-            // HTML error page would reach the hex parser below, producing a
-            // misleading "digest mismatch" log entry instead of the real cause.
-            // The install is blocked either way — this guard exists solely to
-            // name the failure correctly in Console.app.
-            guard let checksumHTTP = checksumResponse as? HTTPURLResponse else {
-                throw URLError(.badServerResponse)
-            }
-            guard checksumHTTP.statusCode == 200 else {
-                appUpdaterLogger.error("checksum sidecar returned HTTP \(checksumHTTP.statusCode, privacy: .public) — release may have been published without a .sha256 file")
-                throw URLError(.badServerResponse)
-            }
-
-            // ── Parse and validate the expected hex string ───────────────────
-            let rawChecksum = String(bytes: checksumData, encoding: .utf8) ?? ""
-            let expectedHex = rawChecksum
-                .trimmingCharacters(in: .whitespacesAndNewlines)
-                .components(separatedBy: .whitespaces).first ?? ""
-
-            // Explicit guard: an empty expectedHex means the sidecar was
-            // zero-length or contained only whitespace (the HTTP 200 / non-200
-            // cases are already handled above). verifyChecksum would throw a
-            // mismatch anyway ("" != 64-char SHA-256 hex), but this guard
-            // prevents a future edit to verifyChecksum from accidentally
-            // treating empty as "skip verification".
-            guard !expectedHex.isEmpty else {
-                appUpdaterLogger.error("checksum sidecar returned HTTP 200 but body was empty or whitespace-only")
-                throw URLError(.cannotDecodeContentData)
-            }
-
-            try await verifyChecksum(zipURL: downloadedURL, expectedHex: expectedHex)
-
-            let destination = try cachedZipDestination(version: version)
-            try? FileManager.default.removeItem(at: destination)
-            try FileManager.default.moveItem(at: downloadedURL, to: destination)
-
-            defaults.set(version, forKey: keys.cachedUpdateVersion)
-            defaults.set(destination.path, forKey: keys.cachedUpdateZipPath)
-
-            state.setDownloadComplete(zipURL: destination, version: version)
-            isDownloading = false
-        } catch {
-            if let tmp = tempURL {
-                try? FileManager.default.removeItem(at: tmp)
-            }
-            // `isDownloading` is cleared on the same @MainActor turn as
-            // `setUpdateFailed()` — no intermediate state is observable. A
-            // subsequent `handle()` call cannot slip through between these two
-            // lines because @MainActor serialises all callers onto one executor.
-            // Do NOT split these or add an `await` between them.
-            isDownloading = false
-            state.setUpdateFailed()
-        }
     }
 }
